requirements.txt
streamlit
numpy
Pillow
joblib
mahotas
scikit-learn
